+++ packages/orq-rc/src/orq_ai_sdk/models/createidentityop.py
@@ -137,7 +137,7 @@ class CreateIdentityResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2026-02-19T09:07:33.562Z")
+    updated: Optional[datetime] = parse_datetime("2026-02-19T11:04:28.599Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")