+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KHTJFKBSHN1K70ZF3D62E8YE"
+        "01KHTS5PY17K6YTVYG7KXEF3KX"
     )
     r"""The unique identifier of the data source"""