+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KHTJFKBY9DGMPM814NXERJ62"
+        "01KHTS5PY5RDS7K109X4N8FQME"
     )
     r"""The unique identifier of the data source"""